backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15.0,<4.0
djangorestframework-simplejwt>=5.3.0,<6.0
channels>=4.1.0,<5.0
django-simple-history>=3.7.0,<4.0
django-cors-headers>=4.3.0,<5.0
python-dotenv>=1.0.0,<2.0
daphne>=4.1.0,<5.0

# Database (opcional - PostgreSQL em produção)
dj-database-url>=2.2.0,<3.0

# API docs
drf-spectacular>=0.27.0,<1.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.8.0,<5.0
pytest-cov>=5.0,<6.0
factory-boy>=3.3.0,<4.0
